+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ releases, in reverse chronological order.
 * VAT conditions in models are now limited to know types -- this should very
   much help create UIs and forms. If you come across a missing VAT condition,
   please open an issue for it.
+* Improved the documentation surrounding PDF generation.
 
 3.2.1
 -----